+++ src/internal/events/events.go
@@ -28,17 +28,13 @@ const (
 	tenantIDDeprecated = "m365_tenant_hash_deprecated"
 
 	// Event Keys
-	RepoInit         = "Repo Init"
-	RepoConnect      = "Repo Connect"
-	RepoUpdate       = "Repo Update"
-	BackupStart      = "Backup Start"
-	BackupEnd        = "Backup End"
-	RestoreStart     = "Restore Start"
-	RestoreEnd       = "Restore End"
-	ExportStart      = "Export Start"
-	ExportEnd        = "Export End"
-	MaintenanceStart = "Maintenance Start"
-	MaintenanceEnd   = "Maintenance End"
+
+	RepoInit       = "Repo Init"
+	RepoConnect    = "Repo Connect"
+	BackupEnd      = "Backup End"
+	RestoreEnd     = "Restore End"
+	ExportEnd      = "Export End"
+	MaintenanceEnd = "Maintenance End"
 
 	// Event Data Keys
 	BackupCreateTime = "backup_creation_time"