+++ src/internal/kopia/upload_test.go
@@ -883,7 +883,7 @@ func makeManifestEntry(
 	var reasons []identity.Reasoner
 
 	for _, c := range categories {
-		reasons = append(reasons, NewReason(tenant, resourceOwner, service, c))
+		reasons = append(reasons, identity.NewReason(tenant, resourceOwner, service, c))
 	}
 
 	return ManifestEntry{